+++ FWI/C/NG_FWI.c
@@ -34,55 +34,58 @@ double mcdc_to_dc(double mcdc) {
   return 400.0 * log(400.0 / mcdc);
 }
 
-double hourly_fine_fuel_moisture(double lastmc, double temp, double rh, double ws,
-  double rain, double time_increment)
-{
-  static const double rf = 42.5;
-  static const double drf = 0.0579;
-  /* use moisture directly instead of converting to/from ffmc */
-  /* expects any rain intercept to already be applied */
-  double mo = lastmc;
-  if (rain != 0.0)
-  {
-    /* duplicated in both formulas, so calculate once */
-    /* lastmc == mo, but use lastmc since mo changes after first equation */
-    mo += rf * rain * exp(-100.0 / (251 - lastmc)) * (1.0 - exp(-6.93 / rain));
-    if (lastmc > 150)
-    {
-      mo += 0.0015 * pow(lastmc - 150, 2) * sqrt(rain);
+double fine_fuel_moisture_code(
+  double mc_0,
+  double temp,
+  double rh,
+  double ws,
+  double prec,
+  double prec_sum,
+  double delta_t
+) {
+  double prec_ffmc, mc_r, e_1, e_w, e_d, eta, k0, k, mc;
+  /*** calculate effective precipitation ***/
+  if (prec_sum + prec <= PREC_MIN_FFMC) {  // not enough rain
+    prec_ffmc = 0.0;
+  }
+  else if (prec_sum > PREC_MIN_FFMC) {  // already saturated canopy before
+    prec_ffmc = prec;
+  }
+  else {  // canopy just saturated this timestep, apply intercept
+    prec_ffmc = prec_sum + prec - PREC_MIN_FFMC;
+  }
+  /*** calculate moisture content after rain ***/
+  if (prec_ffmc > 0.0) {
+    mc_r = mc_0 + (
+      42.5 * prec_ffmc * exp(-100.0/(251.0-mc_0)) * (1.0-exp(-6.93/prec_ffmc))
+    );
+    if (mc_0 > 150.0) {
+      mc_r += 1.5e-3 * pow(mc_0-150.0, 2.0) * sqrt(prec_ffmc);
     }
-    if (mo > 250)
-    {
-      mo = 250;
+    if (mc_r > 250.0) {  // cap fine fuel moisture content at 250%
+      mc_r = 250.0;
     }
   }
-  /* duplicated in both formulas, so calculate once */
-  const double e1 = 0.18 * (21.1 - temp) * (1.0 - (1.0 / exp(0.115 * rh)));
-  const double ed = 0.942 * pow(rh, 0.679) + (11.0 * exp((rh - 100) / 10.0)) + e1;
-  const double ew = 0.618 * pow(rh, 0.753) + (10.0 * exp((rh - 100) / 10.0)) + e1;
-
-  double m;
-  if (mo < ed) {
-    m = ew;
-  } else {
-    m = ed;
+  else {
+    mc_r = mc_0;
   }
-  
-  if (mo != ed)
-  {
-    double a1, k0_or_k1, kd_or_kw;
-    /* these are the same formulas with a different value for a1 */
-    if (mo > ed) {
-      a1 = rh / 100.0;
-    } else {
-      a1 = (100.0 - rh) / 100.0;
-    }
-    k0_or_k1 = 0.424 * (1 - pow(a1, 1.7)) + (0.0694 * sqrt(ws) * (1 - pow(a1, 8)));
-    kd_or_kw = (1.0 / 0.50) * drf * k0_or_k1 * exp(0.0365 * temp);
-    m += (mo - m) * pow(10, -kd_or_kw * time_increment);
-    
+  /*** calculate drying phase (drying or wetting) ***/
+  e_1 = 0.18 * (21.1-temp) * (1.0-exp(-0.115*rh));
+  e_w = 0.618*pow(rh, 0.753) + (10.0*exp((rh-100.0)/10.0)) + e_1;
+  e_d = 0.942*pow(rh, 0.679) + (11.0*exp((rh-100.0)/10.0)) + e_1;
+  if (mc_r < e_d) {
+    eta = (100.0-rh) / 100.0;
+    k0 = 0.424*(1.0-pow(eta, 1.7)) + 0.0694*sqrt(ws)*(1.0-pow(eta, 8.0));
+    k = 0.1158 * exp(0.0365*temp) * k0;
+    mc = e_w + (mc_r-e_w)*pow(10.0, -k*delta_t);
   }
-  return m;
+  else {
+    eta = rh / 100.0;
+    k0 = 0.424*(1.0-pow(eta, 1.7)) + 0.0694*sqrt(ws)*(1.0-pow(eta, 8.0));
+    k = 0.1158 * exp(0.0365*temp) * k0;
+    mc = e_d + (mc_r-e_d)*pow(10.0, -k*delta_t);
+  }
+  return mc;
 }
 
 double duff_moisture_code(double last_mcdmc, int hour,
@@ -92,10 +95,10 @@ double duff_moisture_code(double last_mcdmc, int hour,
   double mr, mcdmc;
 
   // wetting
-  if (prec_cumulative_prev + prec > DMC_INTERCEPT) {  // prec_cumulative above threshold
+  if (prec_cumulative_prev + prec > PREC_MIN_DMC) {  // prec_cumulative above threshold
     double rw, last_dmc, b;
 
-    if (prec_cumulative_prev <= DMC_INTERCEPT) {  // just passed threshold
+    if (prec_cumulative_prev <= PREC_MIN_DMC) {  // just passed threshold
       rw = (prec_cumulative_prev + prec) * 0.92 - 1.27;
     } else {
       rw = prec * 0.92;
@@ -149,10 +152,10 @@ double drought_code(double last_mcdc, int hour, double temp, double prec,
   double mr, mcdc;
 
   // wetting
-  if (prec_cumulative_prev + prec > DC_INTERCEPT) {  // prec_cumulative above threshold
+  if (prec_cumulative_prev + prec > PREC_MIN_DC) {  // prec_cumulative above threshold
     double rw;
 
-    if (prec_cumulative_prev <= DC_INTERCEPT) {  // just passed threshold
+    if (prec_cumulative_prev <= PREC_MIN_DC) {  // just passed threshold
       rw = (prec_cumulative_prev + prec) * 0.83 - 1.27;
     } else {
       rw = prec * 0.83;